refactor: extract schemas in files
