+++ src/Schema/SchemaOptions.ts
@@ -1,0 +1,4 @@
+export interface SchemaOptions<TInput> {
+  default?: TInput;
+  coerce?: boolean;
+}
